Move survey helper functions to utils/src/stages/survey_stage.ts
