+++ frontend/src/components/stages/survey_per_participant_view.ts
@@ -34,13 +34,13 @@ import {
   isMultipleChoiceImageQuestion,
   isQuestionVisible,
   getVisibleSurveyQuestions,
+  isSurveyComplete,
 } from '@deliberation-lab/utils';
 
 import {core} from '../../core/core';
 import {CohortService} from '../../services/cohort.service';
 import {ParticipantService} from '../../services/participant.service';
 import {ParticipantAnswerService} from '../../services/participant.answer';
-import {isSurveyComplete} from '../../shared/stage.utils';
 
 import {styles} from './survey_view.scss';
