+++ frontend/src/components/stages/survey_summary_view.ts
@@ -22,11 +22,9 @@ import {
   TextSurveyAnswer,
   TextSurveyQuestion,
   isMultipleChoiceImageQuestion,
-} from '@deliberation-lab/utils';
-import {
   isSurveyComplete,
   isSurveyAnswerComplete,
-} from '../../shared/stage.utils';
+} from '@deliberation-lab/utils';
 
 import {unsafeHTML} from 'lit/directives/unsafe-html.js';
 import {convertMarkdownToHTML} from '../../shared/utils';
