+++ frontend/src/shared/stage.utils.ts
@@ -1,85 +1,5 @@
-import {
-  ChatMessage,
-  CheckSurveyAnswer,
-  SurveyAnswer,
-  SurveyQuestion,
-  SurveyQuestionKind,
-  TextSurveyAnswer,
-  TextSurveyQuestion,
-  UnifiedTimestamp,
-} from '@deliberation-lab/utils';
+import {ChatMessage, UnifiedTimestamp} from '@deliberation-lab/utils';
 import {Timestamp} from 'firebase/firestore';
-
-/** Returns required questions from survey stage. */
-export function getRequiredSurveyQuestions(questions: SurveyQuestion[]) {
-  return questions.filter(
-    (question) =>
-      !(question.kind === SurveyQuestionKind.CHECK && !question.isRequired),
-  );
-}
-
-/** Returns optional questions from survey stage. */
-export function getOptionalSurveyQuestions(questions: SurveyQuestion[]) {
-  return questions.filter(
-    (question) =>
-      question.kind !== SurveyQuestionKind.CHECK || !question.isRequired,
-  );
-}
-
-/** Checks whether all required questions are answered. */
-export function isSurveyComplete(
-  questions: SurveyQuestion[],
-  answerMap: Record<string, SurveyAnswer>,
-) {
-  const required = getRequiredSurveyQuestions(questions);
-  for (const question of required) {
-    const answer = answerMap[question.id];
-    if (!isSurveyAnswerComplete(answer, question)) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-/** Checks whether given survey answer is complete.
- * Optionally include SurveyQuestion if there are additional validation constraints.
- */
-export function isSurveyAnswerComplete(
-  answer: SurveyAnswer | undefined,
-  question?: SurveyQuestion,
-) {
-  if (!answer) {
-    return false;
-  }
-
-  switch (answer.kind) {
-    case SurveyQuestionKind.CHECK:
-      return (answer as CheckSurveyAnswer).isChecked;
-
-    case SurveyQuestionKind.TEXT:
-      const textAnswer = answer as TextSurveyAnswer;
-      // Check if text exists
-      if (textAnswer.answer.trim() === '') {
-        return false;
-      }
-      // Check character count requirements if question provided
-      if (question && question.kind === SurveyQuestionKind.TEXT) {
-        const textQuestion = question as TextSurveyQuestion;
-        const length = textAnswer.answer.trim().length;
-        if (
-          textQuestion.minCharCount !== undefined &&
-          length < textQuestion.minCharCount
-        ) {
-          return false;
-        }
-      }
-      return true;
-    default:
-      // All other answer types are complete as long as they exist
-      return true;
-  }
-}
 
 /** Returns the timestamp of the first chat message, or null if none. */
 export function getChatStartTimestamp(
